+++ src/Auth.js
@@ -1,6 +1,7 @@
 import l from "./Logger";
 import platform from "platform";
 import EliveError from "./Error";
+
 // import ipinfo from "ipinfo";
 // ipinfo((err, cloc) => {
 //   console.log(cloc.country);
@@ -57,17 +58,19 @@ export default async function Auth(ctx) {
     });
   } catch (e) {
     console.error(e);
-    throw new EliveError(
-      `Auth is failed with id:${ctx.config.credential.serviceId}/ key:${
+    throw new EliveError({
+      code: 1600,
+      text: `Auth is failed with id:${ctx.config.credential.serviceId}/ key:${
         ctx.config.credential.key
       }`
-    );
+    });
   }
   if (!ctx.token)
-    throw new EliveError(
-      `failed to auth with id: ${ctx.config.credential.serviceId} and key: ${
-        ctx.config.credential.key
-      }`
-    );
+    throw new EliveError({
+      code: 1600,
+      text: `failed to auth with id: ${
+        ctx.config.credential.serviceId
+      } and key: ${ctx.config.credential.key}`
+    });
   l.d("success auth");
 }