+++ src/Error.js
@@ -1,7 +1,9 @@
-export default class SlimeError extends Error{
-  constructor(message, e){
-    super(message)
-    this.name='ELiveError'
-    if (e) console.error(e)
+export default class ELiveError extends Error {
+  constructor(message, e) {
+    super(message.text);
+    this.name = "ELiveError";
+    this.code = message.code;
+    this.text = message.text;
+    if (e) console.error(e);
   }
 }