+++ src/SigMsgHandler.js
@@ -75,6 +75,13 @@ function SigMsgHandler(ctx) {
         });
     },
     onCast(msg) {
+      if (msg.status > 3099) {
+        ctx.callEvent({
+          name: "error",
+          param: new EliveError({ code: msg.status, text: msg.desc })
+        });
+        return;
+      }
       l.i(`start onCast with chid: ${msg.channel.id}`);
       ctx.channel = msg.channel;
       ctx.callEvent({ name: "onCast", param: { channel: ctx.channel } });
@@ -175,7 +182,10 @@ function SigMsgHandler(ctx) {
           l.i("remote description is set");
         })
         .catch(e => {
-          throw new EliveError("remote description is wrong");
+          throw new EliveError({
+            code: "1300",
+            text: "remote description is wrong"
+          });
         });
 
       if (desc.type === "offer") {
@@ -198,7 +208,10 @@ function SigMsgHandler(ctx) {
             });
           })
           .catch(e => {
-            throw new EliveError("failed to create answer", e);
+            throw new EliveError(
+              { code: "1300", text: "failed to create answer" },
+              e
+            );
           });
       }
     },
@@ -342,7 +355,7 @@ function SigMsgHandler(ctx) {
       ctx.state = "FAIL";
       ctx.endTime = new Date().getTime();
       l.t(ctx, util.makeTransactionLog(ctx));
-      throw new EliveError("ice connecting is failed");
+      throw new EliveError({ code: "1300", text: "ice connecting is failed" });
     } else if (ctx.peerConnection.iceConnectionState === "disconnected") {
       // 상대 peer에 의해 rtc con이 종료되었을 경우
       if (ctx.state !== "CLOSE") {
@@ -381,7 +394,10 @@ function SigMsgHandler(ctx) {
         ctx.state = "FAIL";
         ctx.endTime = new Date().getTime();
         l.t(ctx, util.makeTransactionLog(ctx));
-        throw new EliveError("ice connecting is failed");
+        throw new EliveError({
+          code: "1300",
+          text: "ice connecting is failed"
+        });
         break;
       case "disconnected": // 상대 peer에 의해 rtc con이 종료되었을 경우
         if (ctx.state !== "CLOSE") {