+++ test/integration/simplecall.js
@@ -1,0 +1,8 @@
+"use strict";
+require(".")({
+  name: "SimpleCall",
+  path: "./test/integration/simplecall",
+  host: "localhost",
+  port: 5000,
+  timeout: 10000
+});