+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+## [3.3.2](https://github.com/elasticlive/web-sdk/compare/v3.3.1...v3.3.2) (2019-03-18)
+
+
+### Bug Fixes
+
+* bug no remoteStream when connect ([7d899ff](https://github.com/elasticlive/web-sdk/commit/7d899ff))
+
+## [3.3.1](https://github.com/elasticlive/web-sdk/compare/v3.3.0...v3.3.1) (2019-03-17)
+
+
+### Bug Fixes
+
+* add integration test code [#1](https://github.com/elasticlive/web-sdk/issues/1) ([0eff891](https://github.com/elasticlive/web-sdk/commit/0eff891))
+
+# [3.3.0](https://github.com/elasticlive/web-sdk/compare/v3.2.0...v3.3.0) (2019-03-11)
+
+
+### Bug Fixes
+
+* All error should include code and text field [#1](https://github.com/elasticlive/web-sdk/issues/1) ([6f19a14](https://github.com/elasticlive/web-sdk/commit/6f19a14))
+* change SlimeError to ELiveError ([e22a8ae](https://github.com/elasticlive/web-sdk/commit/e22a8ae))
+
+
+### Features
+
+* **test:** added config test ([6d683df](https://github.com/elasticlive/web-sdk/commit/6d683df))
+
 # [3.2.0](https://github.com/elasticlive/web-sdk/compare/v3.1.0...v3.2.0) (2019-03-08)